requirements .txt
aniso8601==9.0.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
exceptiongroup==1.1.1
flake8==6.0.0
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
praw==7.7.0
prawcore==2.3.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.6.2
update-checker==0.18.0
urllib3==2.0.2
websocket-client==1.5.2
Werkzeug==2.3.4
zope.interface==6.0
